Encapsulate task handling requests in our own class
There are some properties of the request we want to access from various
places (the Cloud Tasks headers / our parsed `task type` etc) so passing
the raw ServerRequestInterface would give duplication. Instead wrap it
so that handlers can still get at it if required but with extra metadata
and the most common properties surfaced directly.