Rewrite CloudTaskCreator and config to be done by task type, not queue
It makes more sense that our global config is organised by the specific
task types the app understands, and includes the default routing to
handler URLs. That allows for easy:

* Switching between target domains in dev / CI
* Routing specific tasks to an external handler (e.g. cloud run) where
  others are handled by the appserver
* Routing specific tasks to specific queues (e.g. high-priority /
  low-priority) without application-level changes.
* Similarly, configuring the service account to sign individual tasks
  by type.

It's still possible to define task types that we can dispatch but not
process, or that we can process but don't dispatch, but for the more
common cases where tasks are sent and received in the same codebase
it will allow for a lot more convention-over-config cleaning up
of the wiring between tasks.

Also added support for sending task headers and bodies.