Refactor authentication failure results
It's possible that a signature verification will fail due for transient reasons e.g.:

 - time (expired / not yet valid)
 - provider error causing a mismatch between certs/keys and the signed token
 - network error during certificate discovery

And the oidc-token-verifier does not expose the failure reasons in a stable API,
it just provides the upstream exception from firebase/jwt which is itself not always
strongly typed depending on the exact nature of the failure.

Generally in these cases:

a) if it is Cloud Tasks we want them to retry, maybe the next attempt the time will
   be valid / they will generate a new token with valid sig / etc. And anyway it's
   better that this job stays in the queue so we can manually resolve / retry it
   rather than hunting through the logs for ones that had auth failures.

b) if it's not Cloud Tasks it's better to return a semantically-valid 4xx error
   to get the client to go away.

So the code now treats all missing auth token as a `401` and all failures to validate
the token as a `403`.