Relax authentication rules for verifying JWT audience to ignore mismatched protocol or hostname.
In a loadbalanced environment, the `ServerRequestInterface` may not always have accurate information
on whether the request was http/s, or the external hostname that was routed to the appserver.
Therefore only validate the path and querystring (which identifies the operation to be performed).
This still prevents reusing a token to perform a different operation. You should anyway be using
different service accounts for different environments, which in itself protects against using
e.g. a QA token to authorise the same operation in production. Therefore the hostname etc is not
relevant to the authentication/authorisation.